refactor(core): extract sandbox scripts to dedicated package
Move sandbox scripts from @vm0/core to new @vm0/sandbox-scripts package.
This makes @vm0/core truly platform-agnostic with no Node.js dependency.

- Create packages/sandbox-scripts with own tsconfig, build, and lint config
- Move all sandbox script source (14 files), tests (8 files), and build script
- Remove @types/node, esbuild, tsx, memfs from core devDependencies
- Remove "node" from core tsconfig types
- Add "DOM" to core lib for FormData/Blob types used in contracts
- Update e2b-executor.ts to import from @vm0/sandbox-scripts
- Update turbo.json and knip.json for new package

Refs #2721
