+++ turbo/packages/core/eslint.config.mjs
@@ -5,12 +5,4 @@ export default [
   {
     ignores: ["**/dist/**"],
   },
-  {
-    // Sandbox scripts run in E2B/Firecracker VM, not in turbo build system.
-    // Environment variables are injected by the sandbox orchestrator.
-    files: ["**/sandbox/scripts/**/*.ts"],
-    rules: {
-      "turbo/no-undeclared-env-vars": "off",
-    },
-  },
 ];
