+++ turbo/packages/core/src/index.ts
@@ -6,6 +6,5 @@ export * from "./version-id";
 export * from "./storage-names";
 export * from "./github-url";
 export * from "./frameworks";
-export * from "./sandbox/scripts";
 export * from "./feature-switch-key";
 export * from "./feature-switch";
