+++ turbo/packages/sandbox-scripts/eslint.config.mjs
@@ -1,0 +1,16 @@
+import { config } from "@vm0/eslint-config/base";
+
+export default [
+  ...config,
+  {
+    ignores: ["**/dist/**"],
+  },
+  {
+    // This entire package contains sandbox scripts and their tests.
+    // Environment variables are injected by the sandbox orchestrator.
+    files: ["src/**/*.ts"],
+    rules: {
+      "turbo/no-undeclared-env-vars": "off",
+    },
+  },
+];
