+++ turbo/packages/sandbox-scripts/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@vm0/core",
-  "version": "8.6.1",
+  "name": "@vm0/sandbox-scripts",
+  "version": "0.1.0",
   "private": true,
   "type": "module",
   "exports": {
@@ -10,22 +10,22 @@
     }
   },
   "scripts": {
+    "build": "tsx scripts/build-sandbox-scripts.ts",
     "test": "vitest run",
     "test:watch": "vitest",
     "lint": "oxlint && eslint . --max-warnings 0",
     "check-types": "tsc --noEmit"
   },
   "devDependencies": {
+    "@types/node": "^24.3.0",
     "@vm0/eslint-config": "workspace:*",
     "@vm0/typescript-config": "workspace:*",
+    "esbuild": "^0.24.0",
     "eslint": "^9.34.0",
+    "memfs": "^4.56.10",
     "oxlint": "^1.14.0",
+    "tsx": "^4.20.6",
     "typescript": "5.9.2",
     "vitest": "^4.0.18"
-  },
-  "dependencies": {
-    "@ts-rest/core": "3.53.0-rc.1",
-    "yaml": "^2.8.1",
-    "zod": "^4.1.12"
   }
 }
