+++ turbo/packages/sandbox-scripts/src/scripts/lib/secret-masker.ts
