+++ turbo/packages/sandbox-scripts/src/scripts/mock-claude.ts
