+++ turbo/packages/sandbox-scripts/tsconfig.json
@@ -1,9 +1,9 @@
 {
   "extends": "@vm0/typescript-config/base.json",
   "compilerOptions": {
-    "lib": ["ES2022", "DOM"],
-    "types": ["vitest/globals"]
+    "lib": ["ES2022"],
+    "types": ["vitest/globals", "node"]
   },
-  "include": ["src/**/*"],
+  "include": ["src/**/*", "scripts/**/*"],
   "exclude": ["node_modules", "dist"]
 }
